refactor: update remaining import paths
- Update import paths in index.js
- Update import paths in cloudfront/index.entry.js
- Update import paths in vercel/index.entry.js

Finalize import path updates to match new directory structure.
