+++ src/adapters/cloudfront/index.entry.js
@@ -4,18 +4,18 @@
  * @description Main entry point for the Edge Worker
  */
 // CDN specific imports
-import CloudFrontAdapter from './cdn-adapters/cloudfront/cloudFrontAdapter';
-import CloudFrontDynamoDBInterface from './cdn-adapters/cloudfront/cloudFrontDynamoDBInterface';
+import CloudFrontAdapter from './cloudfrontAdapter';
+import CloudFrontDynamoDBInterface from './cloudfrontKVInterface';
 
 // Application specific imports
-import CoreLogic from './coreLogic'; // Assume this is your application logic module
-import OptimizelyProvider from './_optimizely_/optimizelyProvider';
-import defaultSettings from './_config_/defaultSettings';
-import * as optlyHelper from './_helpers_/optimizelyHelper';
-import { getAbstractionHelper } from './_helpers_/abstractionHelper';
-import Logger from './_helpers_/logger';
-import EventListeners from './_event_listeners_/eventListeners';
-import handleRequest from './_api_/apiRouter';
+import CoreLogic from '../../core/providers/coreLogic'; // Assume this is your application logic module
+import OptimizelyProvider from '../../core/providers/optimizelyProvider';
+import defaultSettings from '../../utils/config/defaultSettings';
+import * as optlyHelper from '../../utils/optimizelyHelper';
+import { getAbstractionHelper } from '../../utils/abstractionHelper';
+import Logger from '../../utils/logging/logger';
+import EventListeners from '../../core/providers/events/eventListeners';
+import handleRequest from '../../core/api/apiRouter';
 //
 let abstractionHelper, logger;
 // Define the request, environment, and context objects after initializing the AbstractionHelper
@@ -101,7 +101,7 @@ export async function handler(event, context) {
 	const isGetMethod = httpMethod === 'GET';
 
 	// Check if the request is for the datafile operation
-	const datafileOperation = pathName === '/v1/datafile';
+	const datafileOperation = pathName === '/v1/datafile";
 
 	// Check if the request is for the config operation
 	const configOperation = pathName === '/v1/config';
